--- a/slim.go
+++ b/slim.go
@@ -29,6 +29,16 @@ const (
 	sExpr
 )
 
+var SPACE []byte = []byte(" ")
+
+var NEW_LINE []byte = []byte("\n")
+var LESS_THAN_SLASH []byte = []byte("</")
+var GREATER_THAN_NEW_LINE []byte = []byte(">\n")
+var SLASH_GREATER_THAN_NEW_LINE []byte = []byte ("/>\n")
+var DOUBLE_QUOTE []byte = []byte ("\"")
+var GREATER_THAN []byte = []byte (">")
+var LESS_THAN []byte = []byte ("<")
+
 var emptyElement = []string{
 	"doctype",
 	"area",
@@ -113,6 +123,13 @@ func rubyInline(v *vm.VM, s string) (string, error) {
 	return text, nil
 }
 
+// byteRepeat same as bytes.Repeat but Write to the io.Writer
+func bytesRepeat (out io.Writer, b []byte, count int) {
+	for i :=0; i < count; i++ {
+		out.Write(b)
+	}
+}
+
 func printNode(out io.Writer, v *vm.VM, n *Node, indent int) error {
 	if n.Name == "" && n.Expr == "" {
 		for _, c := range n.Children {
@@ -123,40 +140,53 @@ func printNode(out io.Writer, v *vm.VM, n *Node, indent int) error {
 	} else if n.Name == "/" {
 		return nil
 	} else if n.Name == "/!" {
-		out.Write([]byte(strings.Repeat(" ", indent*2) + "<!-- " + n.Text + " -->\n"))
+		bytesRepeat(out, SPACE, indent * 2)
+		out.Write([]byte("<!-- "))
+		out.Write([]byte(n.Text))
+		out.Write([]byte(" -->\n"))
 	} else {
 		// FIXME
 		doctype := n.Name == "doctype"
 		if doctype {
-			out.Write([]byte(strings.Repeat(" ", indent*2) + "<!" + n.Name + " html"))
+			bytesRepeat(out, SPACE, indent*2)
+			out.Write([]byte("<!"))
+			out.Write([]byte(n.Name))
+			out.Write([]byte(" html"))
 		} else if n.Name != "" {
 			if n.Name[len(n.Name)-1] == ':' {
 				name := n.Name[:len(n.Name)-1]
 				if name == "javascript" {
 					name = "script"
 				}
-				out.Write([]byte(strings.Repeat(" ", indent*2) + "<" + name))
+				bytesRepeat(out, SPACE, indent*2)
+				out.Write(LESS_THAN)
+				out.Write([]byte(name))
 			} else {
-				out.Write([]byte(strings.Repeat(" ", indent*2) + "<" + n.Name))
+				bytesRepeat(out, SPACE, indent*2)
+				out.Write(LESS_THAN)
+				out.Write([]byte(n.Name))
 			}
 		}
 		if n.Id != "" {
-			out.Write([]byte(" id=\"" + n.Id + "\""))
+			out.Write([]byte(" id=\""))
+			out.Write([]byte(n.Id))
+			out.Write(DOUBLE_QUOTE)
 		}
 		if len(n.Class) > 0 {
 			out.Write([]byte(" class=\""))
 			for i, c := range n.Class {
 				if i > 0 {
-					out.Write([]byte(" "))
+					out.Write(SPACE)
 				}
 				out.Write([]byte(c))
 			}
-			out.Write([]byte("\""))
+			out.Write(DOUBLE_QUOTE)
 		}
 		if len(n.Attr) > 0 && !doctype {
 			for _, a := range n.Attr {
 				if a.Value == "" {
-					out.Write([]byte(" " + a.Name))
+					out.Write(SPACE)
+					out.Write([]byte(a.Name))
 				} else {
 					value, err := rubyInline(v, a.Value)
 					if err != nil {
@@ -168,7 +198,7 @@ func printNode(out io.Writer, v *vm.VM, n *Node, indent int) error {
 		}
 		if !isEmptyElement(n.Name) {
 			if n.Name != "" {
-				out.Write([]byte(">"))
+				out.Write(GREATER_THAN)
 			}
 			cr := true
 			if n.Expr != "" {
@@ -191,7 +221,7 @@ func printNode(out io.Writer, v *vm.VM, n *Node, indent int) error {
 						return errors.New("can't iterate: " + n.Expr)
 					}
 					if n.Name != "" {
-						out.Write([]byte("\n"))
+						out.Write(NEW_LINE)
 					}
 					if typ == reflect.Chan {
 						i := 0
@@ -245,7 +275,7 @@ func printNode(out io.Writer, v *vm.VM, n *Node, indent int) error {
 				}
 				out.Write([]byte(text))
 			} else if len(n.Children) > 0 {
-				out.Write([]byte("\n"))
+				out.Write(NEW_LINE)
 				for _, c := range n.Children {
 					if err := printNode(out, v, c, indent+1); err != nil {
 						return err
@@ -264,18 +294,20 @@ func printNode(out io.Writer, v *vm.VM, n *Node, indent int) error {
 				out.Write([]byte(text))
 				cr = false
 			} else if cr {
-				out.Write([]byte("\n"))
+				out.Write(NEW_LINE)
 			}
 			if n.Name != "" {
 				if cr {
-					out.Write([]byte(strings.Repeat(" ", indent*2)))
+					bytesRepeat(out, SPACE, indent*2)
 				}
-				out.Write([]byte("</" + n.Name + ">\n"))
+				out.Write(LESS_THAN_SLASH)
+				out.Write([]byte(n.Name))
+				out.Write(GREATER_THAN_NEW_LINE)
 			}
 		} else if doctype {
-			out.Write([]byte(">\n"))
+			out.Write(GREATER_THAN_NEW_LINE)
 		} else {
-			out.Write([]byte("/>\n"))
+			out.Write(SLASH_GREATER_THAN_NEW_LINE)
 		}
 	}
 	return nil