--- a/slim_test.go
+++ b/slim_test.go
@@ -295,3 +295,20 @@ func TestItem(t *testing.T) {
 		t.Fatalf("expected %v but %v", expect, got)
 	}
 }
+
+func TestComment(t *testing.T) {
+	tmpl, err := ParseFile("testdir/test_comment.slim")
+	if err != nil {
+		t.Fatal(err)
+	}
+	var buf bytes.Buffer
+	err = tmpl.Execute(&buf, nil)
+	if err != nil {
+		t.Fatal(err)
+	}
+	expect := readFile("testdir/test_comment.html")
+	got := buf.String()
+	if expect != got {
+		t.Fatalf("expected %v but %v", expect, got)
+	}
+}