--- a/testdir/test_comment.slim
+++ b/testdir/test_comment.slim
@@ -1,0 +1,1 @@
+/! Comment