Recently I have been working on improving the performance of my go micro services. I want to use Slim to display some health and debug information.

I have created `BenchmarkTemplate_Execute`, based off of `TestSimple`.

Here is the benchcmp output comparing the unchanged code vs. my modifications.

```
benchmark                       old ns/op     new ns/op     delta
BenchmarkTemplate_Execute-8     7508          4893          -34.83%

benchmark                       old allocs     new allocs     delta
BenchmarkTemplate_Execute-8     69             38             -44.93%

benchmark                       old bytes     new bytes     delta
BenchmarkTemplate_Execute-8     3465          3241          -6.46%
``` 

I am happy to explain the changes I have made. I do not think I have affected the architecture of the project with these changes.